Moved prisma to database folder
