+++ src/lib/mocks/Prisma.mock.ts
@@ -1,9 +1,9 @@
 import { PrismaClient } from '@prisma/client';
 import { mockDeep, mockReset, DeepMockProxy } from 'jest-mock-extended';
 
-import prisma from '../prisma';
+import prisma from '../database/client';
 
-jest.mock('../prisma', () => ({
+jest.mock('../database/client', () => ({
   __esModule: true,
   default: mockDeep<PrismaClient>(),
 }));
